requirements.txt
numpy==1.11.1
Pillow==3.3.1
SQLAlchemy==1.0.15
ExifRead==2.1.2
